+++ js/components/CurrentLabels.tsx
@@ -1,8 +1,7 @@
-import {h} from "preact";
+import {h, JSX} from "preact";
 import {AusleseTypes} from "../@types/auslese";
-import {filterDuplicateChoices} from "../lib/helper";
+import {filterDuplicateChoices, generateHierarchicalChoiceTextLabel} from "../lib/helper";
 import {DeleteIcon} from "../lib/icons";
-import JSX = preact.createElement.JSX;
 
 let renderIndex = 0;
 
@@ -16,6 +15,7 @@ export interface CurrentLabelsProps
     onFocus: () => void;
     inputRef: (element: HTMLInputElement) => void;
     disabled?: boolean;
+    includeGroupHeadlineInSelectedChoiceLabel: boolean;
 }
 
 
@@ -27,7 +27,7 @@ export function CurrentLabels (props: CurrentLabelsProps): JSX.Element
         <label class="auslese-current auslese-current-tags" for={randomId}>
             {filterDuplicateChoices(props.choices).map(choice => (
                 <span class="auslese-tag">
-                    <span class="auslese-tag-label">{choice.label}</span>
+                    <span class="auslese-tag-label">{generateHierarchicalChoiceTextLabel(choice, props.includeGroupHeadlineInSelectedChoiceLabel)}</span>
                     {!choice.disabled && (
                         <button type="button" class="auslese-tag-delete" onClick={() => props.onRemove(choice)}>
                             <DeleteIcon />